requirements.txt
astroid==2.2.5
Click==7.0
cycler==0.10.0
dnspython==1.16.0
dominate==2.3.5
Flask==1.0.2
Flask-JSON==0.3.3
gunicorn==19.9.0
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.16.2
opencv-python==4.1.0.25
pafy==0.5.4
pkg-resources==0.0.0
pylint==2.3.1
pymongo==3.7.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.1
six==1.12.0
typed-ast==1.3.1
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
youtube-dl==2019.4.7
